Refactor fixture table logic into dedicated modules
